Crate fast_qr

Example

use fast_qr::{ECL, Version, QRBuilder};

let qrcode = QRBuilder::new("https://example.com/".into())
    .ecl(ECL::H)
    .version(Version::V03)
    .build();
// It is preferable to check qrcode result before
qrcode.unwrap().print();

Example SVG

use fast_qr::{ECL, Version, QRBuilder};
use fast_qr::convert::svg::{SvgBuilder, Shape};

let qrcode = QRBuilder::new("https://example.com/".into())
    .ecl(ECL::H)
    .version(Version::V03)
    .build();

let svg = SvgBuilder::default()
    .shape(Shape::RoundedSquare)
    .to_str(&qrcode.unwrap());

println!("{}", svg);

Re-exports

pub use crate::qr::QRBuilder;

pub use crate::qr::QRCode;

Modules

convert

Converts QRCode matrix to usable

qr

Wrappers to create QRCode

Structs

QROptions

Same as QRBuilder without input.

Enums

ECL

Error Correction Coding has 4 levels

Mask

The different mask patterns. The mask pattern should only be applied to the data and error correction portion of the QR code.

Version

Enum containing all possible QRCode versions

Functions

qr

Generate a QR code from a string. All parameters are automatically set.

qr_opt

Generate a QR code from a string. All parameters are automatically set.